Consolidate checking for valid splits
